+++ CHANGELOG.md
@@ -17,6 +17,7 @@ theme, the application now respects the system preference by default.
 - sensuctl dump can now list the types of supported resources with --types.
 - Added the `sensu_agent_version` field to the `Entity` resource, which reflects
 the Sensu semver version of the agent entity.
+- Added the `APIKey` resource and HTTP API support for POST, GET, PATCH and DELETE.
 
 ### Changed
 - [Web] Github is not always the best place for feature requests and discussion,